Simplify `check_*` functions
The programmer's "default" choice should be to error on an unhandled
status code; the other option is to check for both errors and status
codes at once. Anything else you can handle with `is_*` or a match
statement.

Also threw in a README comment that I am not ready to act on.