Expose Android WSI types and functions
These don't depend on any external headers, so they actually could have
been available all this time.

It would be conceivable to define types in the crate to expose all
platform-specific functionality, but that requires more detailed
knowledge of the platform headers. I think I will trial this on X11 at
least as a demonstration.