Add skeleton code to expose platform-specifics
It should be possible to create an X11 window at this very moment (with
the appropriate Xlib bindings). All platform-specific code could easily
be exposed at this point by copying definitions from the appropriate
headers.

The only thing to keep in mind is that it may require a typecast to
convert to the specific type exported by this crate, e.g. it would
certainly be necessary to convert a pointer obtained from the platform
API to an extern/opaque type pointer as defined in this crate.